+++ packages/cli/src/tokens/build.ts
@@ -1,11 +1,13 @@
 import path from 'node:path';
 import type { ThemeObject } from '@tokens-studio/types';
 import chalk from 'chalk';
+import * as R from 'ramda';
 import type { DesignToken } from 'style-dictionary/types';
 import { cleanDir, mkdir, readFile, writeFile } from '../utils.js';
-import { createThemeCSSFiles } from './format.js';
+import { createThemeCSSFiles, defaultFileHeader } from './process/theme.js';
+
 import { type BuildOptions, processPlatform } from './process/platform.js';
-import type { OutputFile } from './types.js';
+import type { DesignsystemetObject, OutputFile } from './types.js';
 
 async function write(files: OutputFile[], outDir: string, dry?: boolean) {
   for (const { destination, output } of files) {
@@ -25,6 +27,12 @@ export const buildTokens = async (options: Omit<BuildOptions, 'process' | '$them
   const outDir = path.resolve(options.outDir);
   const tokensDir = path.resolve(options.tokensDir);
   const $themes = JSON.parse(await readFile(`${tokensDir}/$themes.json`)) as ThemeObject[];
+  let $designsystemet: DesignsystemetObject | undefined;
+
+  try {
+    const $designsystemetContent = await readFile(`${tokensDir}/$designsystemet.json`);
+    $designsystemet = JSON.parse($designsystemetContent) as DesignsystemetObject;
+  } catch (error) {}
 
   console.log(`\n🏗️ Start building tokens in ${chalk.green(tokensDir)}`);
 
@@ -55,8 +63,12 @@ export const buildTokens = async (options: Omit<BuildOptions, 'process' | '$them
     await write(formatted, outDir, options.dry);
   }
 
+  const fileHeader = R.join('')([
+    defaultFileHeader,
+    $designsystemet ? `\ndesign-tokens: v${$designsystemet.version} ` : '',
+  ]);
   // Write theme CSS files (<theme>.css) to the output directory
-  await write(createThemeCSSFiles(processedBuilds), outDir, options.dry);
+  await write(createThemeCSSFiles({ processedBuilds, fileHeader }), outDir, options.dry);
 
   console.log(`\n✅ Finished building tokens!`);
   return processedBuilds;