+++ packages/cli/src/tokens/create/generators/$designsystemet.ts
@@ -1,0 +1,8 @@
+import pkg from '../../../../package.json' with { type: 'json' };
+
+export function generate$Designsystemet() {
+  return {
+    name: pkg.name,
+    version: pkg.version,
+  };
+}